Refactor to allow logical operations on guards.
The examples do not pass because the desugaring of 'long' transitions into
short ones is not implemented.

The main change is that transitions now have a list of (guard, action)
pairs, and guards may be combined with logical operations. The interpreter
should work when transitions have exactly one (guard, action) pair, but
is not well tested.

We also changed the syntax of return transitions from
  x -> y: return  R := *.m(*,*,*)
to
  x -> y: return R := m[3]
(Methods are identified by name and arity.)