A draft of pretty-printing in a text file.
The instrumenter compiles a property from TOPL and a given Java project.
The result must end up as Java data-structures, but generating Java code
that instantiates those data structures sometimes overflows a limit that
Java bytecode has on method sizes. Thus, it is better to dump the compiled
version in a text file that is very easy to parser and write a small parser
in Java.